requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
supabase>=2.0.0
pdfplumber==0.11.0
google-genai>=1.0.0
python-dotenv>=1.1.0
python-multipart==0.0.9
